renamed "either" file
